+++ src/components/footer/footer.tsx
@@ -1,11 +1,11 @@
 import type { FC, ReactNode } from 'react'
 import styles from './footer.module.css'
 
-export type FooterProps = {
+const Footer = ({
+	children
+}: {
 	children: ReactNode
-}
-
-const Footer: FC<FooterProps> = ({ children }) => {
+}) => {
 	return (
 		<footer className={styles.footer}>
 			<div className={styles.footer__content}>{children}</div>