+++ src/components/image/image.tsx
@@ -1,13 +1,14 @@
-import type { FC } from 'react'
 import styles from './image.module.css'
 
-export type ImageProps = {
+const Image = ({
+	role,
+	src,
+	alt
+}: {
 	role: 'banner' | 'hero' | 'thumbnail' | 'feature' | 'background'
 	src: string
 	alt: string
-}
-
-const Image: FC<ImageProps> = ({ role, src, alt }) => {
+}) => {
 	return <img src={src} alt={alt} className={styles[`image--${role}`]} />
 }