+++ src/components/list-item/list-item.module.d.ts
@@ -1,0 +1,6 @@
+declare module "*/list-item.module.css" {
+const styles: {
+"list-item": string
+}
+export default styles
+}