+++ src/components/list/list.module.d.ts
@@ -1,0 +1,6 @@
+declare module "*/list.module.css" {
+const styles: {
+"list": string
+}
+export default styles
+}