+++ src/components/list/list.tsx
@@ -1,0 +1,8 @@
+import type { ReactNode } from 'react'
+import styles from './list.module.css'
+
+const List = ({ children }: { children: ReactNode }) => {
+	return <ul className={styles.list}>{children}</ul>
+}
+
+export default List