+++ src/components/select-option/select-option.tsx
@@ -1,13 +1,15 @@
 import type { FC, ReactNode } from 'react'
 import styles from './select-option.module.css'
 
-export type SelectOptionProps = {
+const SelectOption = ({
+	value,
+	children,
+	disabled
+}: {
 	value: string
 	children: ReactNode
 	disabled?: boolean
-}
-
-const SelectOption: FC<SelectOptionProps> = ({ value, children, disabled }) => {
+}) => {
 	return (
 		<option value={value} disabled={disabled} className={styles['select-option']}>
 			{children}