+++ src/components/stack/stack.module.d.ts
@@ -1,0 +1,10 @@
+declare module "*/stack.module.css" {
+const styles: {
+"stack": string
+"stack--between": string
+"stack--start": string
+"stack--end": string
+"stack--center": string
+}
+export default styles
+}